refactor: split mask sensitive data mysql and postgresql
